+++ lockdown/tests/views.py
@@ -1,7 +1,11 @@
+import datetime
 from django.http import HttpResponse
 
 from lockdown.decorators import lockdown
 from lockdown.forms import AuthForm
+
+YESTERDAY = datetime.datetime.now()-datetime.timedelta(days=1)
+TOMORROW = datetime.datetime.now()+datetime.timedelta(days=1)
 
 
 def a_view(request):
@@ -24,6 +28,36 @@ def overridden_locked_view(request):
 @lockdown(url_exceptions=(r'^/locked/view/with/exception2/',))
 def locked_view_with_exception(request):
     """View, locked by the decorator with url exceptions"""
+    return HttpResponse('A locked view.')
+
+
+@lockdown(until_date=YESTERDAY)
+def locked_view_until_yesterday(request):
+    """View, locked till yesterday"""
+    return HttpResponse('A locked view.')
+
+
+@lockdown(until_date=TOMORROW)
+def locked_view_until_tomorrow(request):
+    """View, locked till tomorrow"""
+    return HttpResponse('A locked view.')
+
+
+@lockdown(after_date=YESTERDAY)
+def locked_view_after_yesterday(request):
+    """View, locked since yesterday"""
+    return HttpResponse('A locked view.')
+
+
+@lockdown(after_date=TOMORROW)
+def locked_view_after_tomorrow(request):
+    """View, locked starting from tomorrow"""
+    return HttpResponse('A locked view.')
+
+
+@lockdown(until_date=YESTERDAY, after_date=TOMORROW)
+def locked_view_until_and_after(request):
+    """View, only not looked between yesterday and tomorrow"""
     return HttpResponse('A locked view.')